default for byName()
Sometimes when searching there may not be an applicable addon. In that case it will be helpful to have a default. Example:

class X(AddonManager):
     ...

X.byName('NonExistentClass', X) will return X. An alternative is to take a useBase:bool parameter and return cls when it is set to True, but this is not as flexible as the default parameter.